+++ app/src/main/java/org/simple/clinic/di/NetworkModule.kt
@@ -9,7 +9,7 @@ import okhttp3.logging.HttpLoggingInterceptor
 import org.simple.clinic.BuildConfig
 import org.simple.clinic.analytics.NetworkAnalyticsInterceptor
 import org.simple.clinic.overdue.AppointmentCancelReason
-import org.simple.clinic.patient.BusinessId
+import org.simple.clinic.patient.businessid.BusinessId
 import org.simple.clinic.patient.PatientSummaryResult
 import org.simple.clinic.patient.sync.PatientPayload
 import org.simple.clinic.user.LoggedInUserHttpInterceptor
