+++ app/src/main/java/org/simple/clinic/patient/PatientProfile.kt
@@ -1,5 +1,7 @@
 package org.simple.clinic.patient
 
+import org.simple.clinic.patient.businessid.BusinessId
+
 data class PatientProfile(
     val patient: Patient,
     val address: PatientAddress,
