+++ app/src/test/java/org/simple/clinic/patient/PatientRepositoryTest.kt
@@ -27,6 +27,7 @@ import org.simple.clinic.bp.BloodPressureMeasurement
 import org.simple.clinic.bp.PatientToFacilityId
 import org.simple.clinic.facility.FacilityRepository
 import org.simple.clinic.patient.PatientSearchResult.PatientNameAndId
+import org.simple.clinic.patient.businessid.BusinessId
 import org.simple.clinic.patient.filter.SearchPatientByName
 import org.simple.clinic.patient.sync.PatientPayload
 import org.simple.clinic.patient.sync.PatientPhoneNumberPayload
